+++ client/src/components/App.js
@@ -1,0 +1,15 @@
+'use strict'
+
+/**
+ * Dependencies
+ */
+
+import React from 'react'
+
+
+export default class App extends React.Component {
+  render() {
+    return <div>Hello Team Easypeasy!</div>
+  }
+}
+