+++ CHANGELOG.md
@@ -12,14 +12,30 @@ Changes are categorized as follows:
 
 ## [Unreleased]
 
+## 2019-04-05, Version TBD
+    refactor folder structure for SPA
+    deps: react-router-dom ^5.0.0
+    deps(dev): install webpack ^4.29.6
+    deps(dev): install webpack-cli ^3.3.0
+    deps(dev): install webpack-dev-server ^3.2.1
+    deps(dev): install @babel/core ^7.4.3
+    deps(dev): install @babel/plugin-proposal-class-properties ^7.4.0
+    deps(dev): install @babel/plugin-transform-runtime ^7.4.3
+    deps(dev): install @babel/preset-env ^7.4.3
+    deps(dev): install @babel/preset-react ^7.0.0
+    deps(dev): install babel-loader ^8.0.5
+    deps(dev): install html-loader ^0.5.5
+    deps(dev): install html-webpack-plugin ^3.2.0
+
+
 ## YYYY-MM-DD, Version 0.1.0
 
-* 7 minor enhancement:
+* 7 minor enhancements:
 
   * Initialized project
-  * deps: install express 4.16.4
-  * deps: install react 16.8.6
-  * deps: install react-dom 16.8.6
-  * deps: install express-react-views 0.11.0
-  * deps: install ava 1.4.1
-  * deps: install puppeteer 1.14.0
+  * deps: install express ^4.16.4
+  * deps: install react ^16.8.6
+  * deps: install react-dom ^16.8.6
+  * deps: install express-react-views ^0.11.0
+  * deps(dev): install ava ^1.4.1
+  * deps(dev): install puppeteer ^1.14.0